requirements.txt
#--find-links https://download.pytorch.org/whl/torch_stable.html
# torch==1.4.0+cu100
scikit-learn
networkx
sacred
dnspython
nltk
tqdm >=4.60.0
transformers>=4.10.0
pytorch-lightning
torchmetrics
pymongo
pytrec-eval
rdflib
accelerate
#datasets >= 1.8.0
sentencepiece != 0.1.92
protobuf
#torch >= 1.3
pandas
owlready2
gensim >= 4.0.1
lxml==4.6.3
